fix(db): 修正 int4 主键误判 —— 真实缺陷是 Rust 侧 i32 与 baseline 的 bigint 不匹配
## 结论反转
DB review 优先级 5 原写「7 张表用 integer 主键，应 SERIAL→BIGSERIAL」。**该判断被推翻**：
在**全新库**上应用当前 baseline，7 张表的 `id` **本来就是 `bigint`**（PK 类型普查：
bigint 181 / text 45 / uuid 3 / varchar 1，`integer` = **0**），DDL 侧**无需改动**。

错因：那个"7 个 int4 PK"是从**过期的本地库** `synapse_test` 读出来的（旧 baseline 建的），
而我当时没有像对其它优先级那样在全新库上复核。**权威只有一个：对当前 baseline 做全新应用。**

## 真实缺陷在相反方向（已修）
baseline 是 `bigint`，Rust 侧却是 `i32`；sqlx 不做 INT8→i32 隐式转换 ⇒ **全新部署运行期解码失败**，
而过期的本地库（int4）恰好让测试全绿 —— 典型的"本地绿、上生产挂"。已宽到 `i64` 对齐：
- `RoomTag.id`（`SELECT id, ... FROM room_tags` 直接解码该列）
- `SearchIndexCursor.id`（游标 `created_ts|id`，id 超 2^31 会在 parse 时溢出）
- 连带测试夹具：`test_mocks/room_tag.rs` 的 `next_id`、`search_index.rs` 里的 `i32::MAX`

## 验证
  cargo check --workspace --all-targets --all-features = **0 error**
      （注意：只跑 `--all-features` 不含 `--all-targets` 会漏掉测试代码——本轮就漏过一次）
  cargo test --test unit --features test-utils = **1815 passed / 0 failed**
  cargo test -p synapse-storage --lib 的 `room_tag::db_tests`：按 baseline 重建库后**由红转绿**
      （用过期 int4 库时反而红——正是"类型已对齐"的直接证据）

## 顺带发现（未修，需裁定，详见报告 §13.4）
1. `widgets.room_id`：baseline 为 **NOT NULL**，但 `Widget.room_id` 是 `Option<String>` 且
   `widget::db_tests` 插入 NULL ⇒ schema/代码/测试三者矛盾（重建库后暴露，与本次改动无关）。
   §6 把该列列为"可空外键"同样是过期库读数。
2. **HEAD 的 fmt 门禁当前是红的**：`synapse-web/src/routes/derived_route_table_always.inc.rs`
   （302KB 生成文件）未格式化 —— 被 B4-5d 给 fmt 棘轮扩扫描面后才暴露。修法：生成器对它跑
   rustfmt，或在棘轮里排除生成的 `*.inc.rs`。

+++ synapse-storage/src/room_tag/mod.rs
@@ -7,7 +7,11 @@ use synapse_common::current_timestamp_millis;
 #[derive(Debug, Clone, Serialize, Deserialize, sqlx::FromRow)]
 pub struct RoomTag {
     /// The `id` field.
-    pub id: i32,
+    ///
+    /// `bigint` in the baseline schema (`room_tags.id BIGSERIAL`): an `i32` here
+    /// decodes fine against a legacy int4 schema but fails at runtime on a fresh
+    /// v12 deploy (sqlx does not coerce INT8 → i32).
+    pub id: i64,
     /// The `user_id` field.
     pub user_id: String,
     /// The `room_id` field.
